+++ tests/test_tikz_path_exporter.py
@@ -4,14 +4,15 @@ import unittest
 
 import sys
 import os
-import lxml
+from io import StringIO
 
 # Use local svg2tikz version
 sys.path.insert(0, os.path.dirname(os.path.realpath(__file__)) + "/../")
 
 # pylint: disable=wrong-import-position
-from svg2tikz.extensions.tikz_export import TikZPathExporter, GraphicsState
+from svg2tikz.extensions.tikz_export import TikZPathExporter
 from tests.common import SVG_4_RECT, SVG_EMPTY, SVG_TEXT
+from inkex.transforms import Vector2d
 
 
 class TestTikZPathExporter(unittest.TestCase):
@@ -34,12 +35,90 @@ class TestTikZPathExporter(unittest.TestCase):
         }
 
         for inp, val_i in units.items():
-            tzpe.options.input_unit = inp
             for out, val_o in units.items():
                 tzpe.options.output_unit = out
-                conv = tzpe.convert_unit(1)
+                conv = tzpe.convert_unit(f"{val_o}{inp}")
+                self.assertTrue(abs(conv - val_i) < 1e-5)
 
-                self.assertEqual(conv, val_i / val_o)
+    def test_convert_unit_coord(self):
+        """Test converting between unit coordinate"""
+        tzpe = TikZPathExporter(inkscape_mode=False)
+        tzpe.height = 5
+        tzpe.options.output_unit = "px"
+        tzpe.options.noreversey = True
+
+        coord = Vector2d(5, 5)
+        output_coord = tzpe.convert_unit_coord(coord)
+        self.assertTupleEqual((coord.x, coord.y), (output_coord.x, output_coord.y))
+
+        tzpe.options.noreversey = False
+        output_coord = tzpe.convert_unit_coord(coord, False)
+        self.assertTupleEqual((coord.x, coord.y), (output_coord.x, output_coord.y))
+
+        output_coord = tzpe.convert_unit_coord(coord, True)
+        self.assertTupleEqual(
+            (coord.x, tzpe.height - coord.y), (output_coord.x, output_coord.y)
+        )
+
+    def test_convert_unit_coords(self):
+        """Test converting between unit coordinates"""
+        tzpe = TikZPathExporter(inkscape_mode=False)
+        tzpe.height = 5
+        tzpe.options.output_unit = "px"
+        tzpe.options.noreversey = True
+
+        coords = [Vector2d(1, 1), Vector2d(2, 2), Vector2d(3, 3)]
+        output_coords = tzpe.convert_unit_coords(coords)
+        for coord, output_coord in zip(coords, output_coords):
+            self.assertTupleEqual((coord.x, coord.y), (output_coord.x, output_coord.y))
+
+        tzpe.options.noreversey = False
+        output_coords = tzpe.convert_unit_coords(coords, False)
+        for coord, output_coord in zip(coords, output_coords):
+            self.assertTupleEqual((coord.x, coord.y), (output_coord.x, output_coord.y))
+
+        output_coords = tzpe.convert_unit_coords(coords, True)
+        for coord, output_coord in zip(coords, output_coords):
+            self.assertTupleEqual(
+                (coord.x, tzpe.height - coord.y), (output_coord.x, output_coord.y)
+            )
+
+    def test_round_value(self):
+        """Test rounding a value"""
+        tzpe = TikZPathExporter(inkscape_mode=False)
+        number = 0.123456789
+        tzpe.options.round_number = 1
+        self.assertEqual(tzpe.round_value(number), 0.1)
+        tzpe.options.round_number = 2
+        self.assertEqual(tzpe.round_value(number), 0.12)
+        tzpe.options.round_number = 3
+        self.assertEqual(tzpe.round_value(number), 0.123)
+
+    def test_round_coord(self):
+        """Test rounding a coordinate"""
+        tzpe = TikZPathExporter(inkscape_mode=False)
+        coord = Vector2d(0.123456789, 0.123456789)
+
+        tzpe.options.round_number = 1
+        output_coord = tzpe.round_coord(coord)
+        self.assertEqual(output_coord.x, 0.1)
+
+    def test_round_coords(self):
+        """Test rounding a coordinates"""
+        tzpe = TikZPathExporter(inkscape_mode=False)
+        coord = Vector2d(0.123456789, 0.123456789)
+        coords = [coord, coord]
+
+        tzpe.options.round_number = 1
+        output_coords = tzpe.round_coords(coords)
+        self.assertEqual(output_coords[1].x, 0.1)
+
+    def test_coord_to_tz(self):
+        """Test rounding and converting a coordinate to tz format"""
+        tzpe = TikZPathExporter(inkscape_mode=False)
+        coord = Vector2d(0.123456789, 0.123456789)
+        tzpe.options.round_number = 1
+        self.assertEqual(tzpe.coord_to_tz(coord), "(0.1, 0.1)")
 
     def test_height(self):
         """Test converting between units"""
@@ -61,98 +140,39 @@ class TestTikZPathExporter(unittest.TestCase):
                 test_val = tzpe.update_height(j)
                 self.assertEqual(true_val, test_val)
 
-    def test_parse(self):
-        """Test parsing file"""
-        tzpe = TikZPathExporter(inkscape_mode=False)
-        path_file = "./tests/testfiles/unicode.svg"
-        tzpe.parse(path_file)
-
-        path_file = "./tests/testfiles/not_exist.svg"
-        self.assertRaises(lxml.etree.XMLSyntaxError, tzpe.parse, path_file)
-
-    def test_get_selected(self):
-        """Test converting between units"""
-        tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_4_RECT)
-        tzpe.options.ids = []
-        tzpe.get_selected()
-        self.assertEqual([], tzpe.selected_sorted)  # Empty
-        ids = ["rect1", "rect2", "rect3"]
-
-        for a in ids:
-            tzpe.options.ids = [a]
-
-            tzpe.get_selected()
-            for elem in tzpe.selected_sorted:
-                test_id = elem.get("id", "")
-                self.assertEqual(test_id, a)  # one elem
-
-        for a in ids:
-            tzpe.options.ids = ["rect4", a]
-
-            tzpe.get_selected()
-            for idx, elem in enumerate(tzpe.selected_sorted):
-                test_id = elem.get("id", "")
-                # Order define in the SVG
-                if idx == 0:
-                    self.assertEqual(test_id, a)  # first elem
-                else:
-                    self.assertEqual(test_id, "rect4")  # second elem
-
-    def test_get_node_form_id(self):
-        """Test converting between units"""
-        tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_4_RECT)
-
-        self.assertEqual(None, tzpe.get_node_from_id("Not_an_id"))
-        ids = ["rect1", "rect2", "rect3"]
-
-        for a in ids:
-            self.assertEqual(a, tzpe.get_node_from_id(a).get("id", ""))
-
     def test_get_color(self):
         """Test getting color"""
         tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_4_RECT)
-        self.assertEqual({}, tzpe.colors)
-        self.assertEqual("red", tzpe.get_color("red"))
-        self.assertEqual("black", tzpe.get_color("r"))  # r is not a valid color
+        tzpe.convert(StringIO(SVG_4_RECT), no_output=True, returnstring=True)
+        self.assertEqual(["navy"], tzpe.colors)
+        # self.assertEqual("red", tzpe.get_color("red"))
+        # self.assertEqual("black", tzpe.get_color("r"))  # r is not a valid color
 
         # It should not be added to list of color
-        self.assertEqual(
-            {
-                "red": "red",
-            },
-            tzpe.colors,
-        )
+        # self.assertEqual(
+        # {
+        # "red": "red",
+        # },
+        # tzpe.colors,
+        # )
 
-        self.assertEqual(
-            "cffffff", tzpe.get_color("rgb(255,255,255)")
-        )  # r is not a valid color
-        self.assertEqual({"red": "red", "rgb(255,255,255)": "cffffff"}, tzpe.colors)
-
-    def test_convert_svgstate_to_tikz(self):
-        """Test the conversion bteween a node style as a list of TikZ options"""
-        tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_4_RECT)
-        gs = GraphicsState(None)
-        node = tzpe.get_node_from_id("rect1")
-        ops, transform = tzpe.convert_svgstate_to_tikz(gs, gs, node)
-        self.assertEqual(["fill"], ops)
-        self.assertEqual([], transform)
+        # self.assertEqual(
+        # "cffffff", tzpe.get_color("rgb(255,255,255)")
+        # )  # r is not a valid color
+        # self.assertEqual({"red": "red", "rgb(255,255,255)": "cffffff"}, tzpe.colors)
 
     def test_get_text(self):
         """Return content of a text node as string"""
         tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_TEXT)
-        test_text = tzpe.get_text(tzpe.get_node_from_id("textNode"))
+        tzpe.convert(StringIO(SVG_TEXT), no_output=True, returnstring=True)
+        test_text = tzpe.get_text(tzpe.svg.getElementById("textNode"))
         true_text = "Test Text\n"
         self.assertEqual(true_text, test_text)
 
     def test_effect(self):
         """Test effect function"""
         tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_EMPTY)
+        tzpe.convert(StringIO(SVG_EMPTY), no_output=True, returnstring=True)
         tzpe.options.output_unit = "mm"
         tzpe.options.input_unit = "cm"
         tzpe.options.noreversey = False
@@ -185,7 +205,7 @@ class TestTikZPathExporter(unittest.TestCase):
     def test_save_raw(self):
         """Test raw saving"""
         tzpe = TikZPathExporter(inkscape_mode=False)
-        tzpe.parse(SVG_4_RECT)
+        tzpe.convert(StringIO(SVG_4_RECT), no_output=True, returnstring=True)
         tzpe.output_code = "Test save"
         tzpe.options.clipboard = False
         tzpe.options.mode = "effect"
@@ -206,27 +226,24 @@ class TestTikZPathExporter(unittest.TestCase):
 \usepackage{tikz}
 
 \begin{document}
-\definecolor{blue}{RGB}{0,0,255}
 \definecolor{navy}{RGB}{0,0,128}
-\definecolor{yellow}{RGB}{255,255,0}
-\definecolor{green}{RGB}{0,128,0}
 
 
 \def \globalscale {1.000000}
 \begin{tikzpicture}[y=1cm, x=1cm, yscale=\globalscale,xscale=\globalscale, inner sep=0pt, outer sep=0pt]
-\path[draw=blue,line width=0.200cm] (0.1, 3.9) rectangle (119.9, -35.900000000000006);
+  \path[draw=blue,line width=0.02cm] (0.01, 3.99) rectangle (11.99, 0.01);
 
 
 
-\path[draw=navy,fill=yellow,line width=1.000cm] (40.00000000000001, -6.000000000000002) rectangle (80.00000000000001, -26.000000000000007);
+  \path[draw=navy,fill=yellow,line width=0.1cm] (4.0, 3.0) rectangle (8.0, 1.0);
 
 
 
-\path[draw=green,line width=1.000cm] (40.00000000000001, -6.000000000000002) rectangle (80.00000000000001, -26.000000000000007);
+  \path[draw=green,line width=0.1cm] (4.0, 3.0) rectangle (8.0, 1.0);
 
 
 
-\path[draw=green,line width=1.000cm] (40.00000000000001, -6.000000000000002) rectangle (80.00000000000001, -26.000000000000007);
+  \path[draw=green,line width=0.1cm] (4.0, 3.0) rectangle (8.0, 1.0);
 
 
 
@@ -234,7 +251,9 @@ class TestTikZPathExporter(unittest.TestCase):
 \end{tikzpicture}
 \end{document}
 """
-        test_path = tzpe.convert(SVG_4_RECT)
+        test_path = tzpe.convert(
+            StringIO(SVG_4_RECT), no_output=True, returnstring=True
+        )
         self.assertEqual(test_path, true_path)